Separate ConsoleWriter.cs
